% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nipter_classes.R
\name{CombinedControlGroup}
\alias{CombinedControlGroup}
\title{NIPTeR control group for CombinedStrands samples}
\usage{
CombinedControlGroup(
  samples = list(),
  description = "General control group",
  sample_sex = NULL,
  sex_source = NULL,
  .cache = new.env(parent = emptyenv())
)
}
\description{
NIPTeR control group for CombinedStrands samples
}
