% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohort.R
\docType{data}
\name{COMPRESSED_BINSIZE}
\alias{COMPRESSED_BINSIZE}
\title{Compressed bin size for synthetic cohort BAMs}
\format{
An object of class \code{integer} of length 1.
}
\usage{
COMPRESSED_BINSIZE
}
\description{
Each 100kb genomic bin is compressed to 100bp, yielding identical bin
counts to GRCh37 at 100kb resolution.
}
\keyword{datasets}
