% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nipter_classes.R
\name{control_names}
\alias{control_names}
\title{Sample names of a control group}
\usage{
control_names(cg, ...)
}
\arguments{
\item{cg}{A \code{NIPTControlGroup} object.}

\item{...}{Reserved for S7 method dispatch; currently unused.}
}
\value{
A character vector.
}
\description{
Sample names of a control group
}
