% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rwisecondorx_predict.R
\name{.apply_blacklist}
\alias{.apply_blacklist}
\title{Apply a blacklist BED to zero out regions}
\usage{
.apply_blacklist(blacklist_path, results_r, results_z, results_w, binsize)
}
\description{
Apply a blacklist BED to zero out regions
}
\keyword{internal}
