% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rwisecondorx_predict.R
\name{.call_aberrations}
\alias{.call_aberrations}
\title{Call aberrations from CBS segments}
\usage{
.call_aberrations(results_c, zscore_cutoff, beta, ref_gender)
}
\description{
Call aberrations from CBS segments
}
\keyword{internal}
