% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rwisecondorx_predict.R
\name{.coverage_normalize_and_mask}
\alias{.coverage_normalize_and_mask}
\title{Coverage normalization and masking}
\usage{
.coverage_normalize_and_mask(sample, ref, ap)
}
\description{
Coverage normalization and masking
}
\keyword{internal}
