% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rwisecondorx_predict.R
\name{.get_optimal_cutoff}
\alias{.get_optimal_cutoff}
\title{Iterative optimal distance cutoff}
\usage{
.get_optimal_cutoff(ref, repeats, sd_multiplier = 3)
}
\description{
Iterative optimal distance cutoff
}
\keyword{internal}
