% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rwisecondorx_utils.R
\name{.gmm_density}
\alias{.gmm_density}
\title{Evaluate the density of a mclust GMM}
\usage{
.gmm_density(fit, x)
}
\arguments{
\item{fit}{mclust model object.}

\item{x}{Numeric vector of evaluation points.}
}
\value{
Numeric vector of density values.
}
\description{
Evaluate the density of a mclust GMM
}
\keyword{internal}
