% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rwisecondorx_output.R
\name{.idx_to_chr_name}
\alias{.idx_to_chr_name}
\title{Convert 1-based chromosome index to display name}
\usage{
.idx_to_chr_name(idx)
}
\description{
Convert 1-based chromosome index to display name
}
\keyword{internal}
