% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rwisecondorx_predict.R
\name{.inflate_results}
\alias{.inflate_results}
\title{Inflate (un-mask) a compressed result vector}
\usage{
.inflate_results(results, mask)
}
\description{
Inflate (un-mask) a compressed result vector
}
\keyword{internal}
