% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rwisecondorx_utils.R
\name{.mclust_gender_fit}
\alias{.mclust_gender_fit}
\title{Fit a 2-component GMM using mclust (gender model)}
\usage{
.mclust_gender_fit(y_fractions)
}
\arguments{
\item{y_fractions}{Numeric vector of Y-chromosome fractions.}
}
\value{
mclust model object.
}
\description{
Fit a 2-component GMM using mclust (gender model)
}
\keyword{internal}
