% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rwisecondorx_predict.R
\name{.normalize}
\alias{.normalize}
\title{Run the full normalization chain for one gender partition}
\usage{
.normalize(
  sample,
  ref,
  ref_gender,
  maskrepeats,
  optimal_cutoff_sd_multiplier = 3,
  within_sample_mask_iterations = 3L,
  within_sample_mask_quantile = 0.99
)
}
\arguments{
\item{sample}{Named list of bin counts.}

\item{ref}{Reference list.}

\item{ref_gender}{\code{"A"}, \code{"F"}, or \code{"M"}.}

\item{maskrepeats}{Number of optimal cutoff iterations.}
}
\value{
List with results_r, results_z, results_w, ref_sizes, m_lr, m_z.
}
\description{
Run the full normalization chain for one gender partition
}
\keyword{internal}
