% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rwisecondorx_predict.R
\name{.normalize_once}
\alias{.normalize_once}
\title{Single within-sample normalization pass}
\usage{
.normalize_once(test_data, test_copy, ref, optimal_cutoff, ct, cp, ap)
}
\description{
Single within-sample normalization pass
}
\keyword{internal}
