% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rwisecondorx_predict.R
\name{.normalize_repeat}
\alias{.normalize_repeat}
\title{Within-sample normalization with iterative aberration masking (3 iterations)}
\usage{
.normalize_repeat(
  test_data,
  ref,
  optimal_cutoff,
  ct,
  cp,
  ap,
  iterations = 3L,
  mask_quantile = 0.99
)
}
\description{
Within-sample normalization with iterative aberration masking (3 iterations)
}
\keyword{internal}
