% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rwisecondorx_predict.R
\name{.split_by_chr}
\alias{.split_by_chr}
\title{Split a full-genome vector into per-chromosome lists}
\usage{
.split_by_chr(data, bins_per_chr)
}
\description{
Split a full-genome vector into per-chromosome lists
}
\keyword{internal}
