% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rwisecondorx_predict.R
\name{.split_nr_by_chr}
\alias{.split_nr_by_chr}
\title{Split a null-ratio matrix into per-chromosome list of matrices}
\usage{
.split_nr_by_chr(nr_matrix, bins_per_chr)
}
\description{
Split a null-ratio matrix into per-chromosome list of matrices
}
\keyword{internal}
