% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rwisecondorx_output.R
\name{.write_aberrations_bed}
\alias{.write_aberrations_bed}
\title{Write aberrations BED file}
\usage{
.write_aberrations_bed(prediction, outprefix)
}
\description{
Write aberrations BED file
}
\keyword{internal}
