% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rwisecondorx_output.R
\name{.write_prediction_output}
\alias{.write_prediction_output}
\title{Internal output writer}
\usage{
.write_prediction_output(prediction, outprefix)
}
\description{
Internal output writer
}
\keyword{internal}
