% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nipter_classes.R
\name{fractions_for_regression}
\alias{fractions_for_regression}
\title{Extract fractions suitable for regression}
\usage{
fractions_for_regression(cg, ...)
}
\arguments{
\item{cg}{A \code{NIPTControlGroup} object.}
}
\value{
A numeric matrix.
}
\description{
For \code{CombinedControlGroup}: 22 x N matrix.
For \code{SeparatedControlGroup}: 44 x N matrix (rows "1F".."22F","1R".."22R").
}
