% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nipter_classes.R
\name{n_bins}
\alias{n_bins}
\title{Number of bins in a NIPTSample}
\usage{
n_bins(x, ...)
}
\arguments{
\item{x}{A \code{NIPTSample} object.}

\item{...}{Reserved for S7 method dispatch; currently unused.}
}
\value{
A positive integer.
}
\description{
Number of bins in a NIPTSample
}
