% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nipter_classes.R
\name{NCVTemplate}
\alias{NCVTemplate}
\title{Pre-computed NCV denominator template}
\usage{
NCVTemplate(
  focus_chromosome = integer(0),
  denominators = character(0),
  ctrl_mean = numeric(0),
  ctrl_sd = numeric(0),
  ctrl_cv = numeric(0),
  shapiro_p = numeric(0),
  test_z_scores = numeric(0),
  test_sample_names = character(0),
  train_sample_names = character(0)
)
}
\description{
Returned by \code{nipter_prepare_ncv()} and consumed by
\code{nipter_ncv_score()}.
}
