% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nipter_classes.R
\name{NIPTControlGroup}
\alias{NIPTControlGroup}
\title{Abstract base class for NIPTeR control groups}
\usage{
NIPTControlGroup(
  samples = list(),
  description = "General control group",
  sample_sex = NULL,
  sex_source = NULL,
  .cache = new.env(parent = emptyenv())
)
}
\description{
Abstract base class for NIPTeR control groups
}
