% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nipter_classes.R
\name{NIPTCorrectionPath}
\alias{NIPTCorrectionPath}
\title{Ordered correction path for autosomal or sex-chromosome processing}
\usage{
NIPTCorrectionPath(steps = list())
}
\arguments{
\item{steps}{List of \code{NIPTCorrectionStep} objects. Defaults to a single
\code{"raw"} step.}
}
\description{
Ordered correction path for autosomal or sex-chromosome processing
}
