% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nipter_classes.R
\name{NIPTCorrectionStep}
\alias{NIPTCorrectionStep}
\title{Typed correction step in a NIPTeR correction path}
\usage{
NIPTCorrectionStep(code = character(0))
}
\arguments{
\item{code}{Correction step code. One of \code{"raw"}, \code{"gc"}, or
\code{"chi"}.}
}
\description{
Typed correction step in a NIPTeR correction path
}
