% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nipter_plot.R
\name{nipter_plot_qc_samples}
\alias{nipter_plot_qc_samples}
\title{Plot sample-level NIPTeR control-group QC}
\usage{
nipter_plot_qc_samples(qc)
}
\arguments{
\item{qc}{A \code{NIPTControlGroupQC} from
\code{\link{nipter_control_group_qc}}.}
}
\value{
A \code{ggplot} object.
}
\description{
Visualises the control-group matching summary from a
\code{\link{NIPTControlGroupQC}} object.
}
