% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nipter_classes.R
\name{NIPTeRSexModel}
\alias{NIPTeRSexModel}
\title{NIPTeR sex-prediction model}
\usage{
NIPTeRSexModel(.data)
}
\arguments{
\item{.data}{Named list payload for the typed constructor.}
}
\description{
A list-like S7 object wrapping one trained sex-classification model.
}
