% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nipter_classes.R
\name{SeparatedControlGroup}
\alias{SeparatedControlGroup}
\title{NIPTeR control group for SeparatedStrands samples}
\usage{
SeparatedControlGroup(
  samples = list(),
  description = "General control group",
  sample_sex = NULL,
  sex_source = NULL,
  .cache = new.env(parent = emptyenv())
)
}
\description{
NIPTeR control group for SeparatedStrands samples
}
