% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nipter_classes.R
\name{strand_type}
\alias{strand_type}
\title{Return the strand type of a NIPTSample}
\usage{
strand_type(x, ...)
}
\arguments{
\item{x}{A \code{NIPTSample} object.}
}
\value{
\code{"combined"} or \code{"separated"}.
}
\description{
Return the strand type of a NIPTSample
}
