% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rwisecondorx_classes.R
\name{WisecondorXObject}
\alias{WisecondorXObject}
\title{Abstract base class for WisecondorX list-like objects}
\usage{
WisecondorXObject(.data)
}
\description{
Abstract base class for WisecondorX list-like objects
}
