% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rwisecondorx_classes.R
\name{WisecondorXPrediction}
\alias{WisecondorXPrediction}
\title{WisecondorX prediction result}
\usage{
WisecondorXPrediction(.data)
}
\description{
A list-like S7 object produced by \code{\link[=rwisecondorx_predict]{rwisecondorx_predict()}}.
}
