% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rwisecondorx_classes.R
\name{WisecondorXReferenceQC}
\alias{WisecondorXReferenceQC}
\title{WisecondorX reference QC report}
\usage{
WisecondorXReferenceQC(.data)
}
\description{
A structured QC report returned by \code{\link[=rwisecondorx_ref_qc]{rwisecondorx_ref_qc()}}.
}
