% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rwisecondorx_output.R
\name{write_wisecondorx_output}
\alias{write_wisecondorx_output}
\title{Write prediction output files}
\usage{
write_wisecondorx_output(prediction, outprefix)
}
\arguments{
\item{prediction}{A \code{WisecondorXPrediction} object.}

\item{outprefix}{Path prefix for output files.}
}
\value{
\code{outprefix} (invisibly).
}
\description{
Writes bins BED, segments BED, aberrations BED, and statistics text file.
}
