% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nipter_classes.R
\name{CombinedStrandsSample}
\alias{CombinedStrandsSample}
\title{NIPTeR sample with combined forward+reverse read counts}
\usage{
CombinedStrandsSample(
  sample_name = character(0),
  binsize = integer(0),
  chrom_lengths = NULL,
  correction = NIPTCorrectionRecord(),
  auto_matrix = NULL,
  sex_matrix_ = NULL
)
}
\arguments{
\item{sample_name}{Sample identifier.}

\item{binsize}{Positive integer bin width in base pairs.}

\item{correction}{A \code{NIPTCorrectionRecord} describing the applied
corrections.}

\item{auto_matrix}{22 x n_bins numeric matrix for autosomes.}

\item{sex_matrix_}{2 x n_bins numeric matrix for X/Y.}
}
\description{
NIPTeR sample with combined forward+reverse read counts
}
\section{Slots}{

\describe{
\item{\code{auto_matrix}}{22 x n_bins numeric matrix (rows = chr 1-22).}

\item{\code{sex_matrix_}}{2 x n_bins numeric matrix (row 1 = X, row 2 = Y).}
}}

