% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rwisecondorx_cbs.R
\name{.compute_statistics}
\alias{.compute_statistics}
\title{Compute per-chromosome statistics}
\usage{
.compute_statistics(
  results_r,
  results_w,
  results_c,
  results_nr,
  bins_per_chr,
  binsize,
  ref_gender,
  gender,
  n_reads
)
}
\description{
Mirrors \code{predict_output._generate_chr_statistics_file()}.
}
\keyword{internal}
