% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nipter_classes.R
\name{n_controls}
\alias{n_controls}
\title{Number of samples in a control group}
\usage{
n_controls(cg, ...)
}
\arguments{
\item{cg}{A \code{NIPTControlGroup} object.}
}
\value{
An integer.
}
\description{
Number of samples in a control group
}
