% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nipter_classes.R
\name{NIPTCorrectionRecord}
\alias{NIPTCorrectionRecord}
\title{Correction record for a NIPTeR sample}
\usage{
NIPTCorrectionRecord(autosomal = "Uncorrected", sex = "Uncorrected")
}
\description{
Tracks the sequence of corrections applied to the autosomal and sex
chromosome matrices of a \code{NIPTSample} object.
}
