% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nipter_classes.R
\name{SeparatedStrandsSample}
\alias{SeparatedStrandsSample}
\title{NIPTeR sample with separate forward and reverse read counts}
\usage{
SeparatedStrandsSample(
  sample_name = character(0),
  binsize = integer(0),
  correction = NIPTCorrectionRecord(),
  auto_fwd = NULL,
  auto_rev = NULL,
  sex_fwd = NULL,
  sex_rev = NULL
)
}
\description{
NIPTeR sample with separate forward and reverse read counts
}
\section{Slots}{

\describe{
\item{\code{auto_fwd}}{22 x n_bins numeric matrix (forward strand).}

\item{\code{auto_rev}}{22 x n_bins numeric matrix (reverse strand).}

\item{\code{sex_fwd}}{2 x n_bins numeric matrix (forward strand; row1=X, row2=Y).}

\item{\code{sex_rev}}{2 x n_bins numeric matrix (reverse strand).}
}}

